+++ src/graph/sampling/randomwalks/get_node_types_cpu.cc
